# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "High-performance AI gateway routing requests to multiple LLM providers (OpenAI, Anthropic, Gemini, Groq, OpenRouter, DeepSeek, Z.ai, xAI, MiniMax, Oracle, Ollama). Drop-in OpenAI-compatible API.",
    "title": "GoModel API",
    "contact": {},
    "version": "1.0"
  },
  "paths": {
    "/admin/api/v1/audit/conversation": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get conversation thread around an audit log entry",
        "parameters": [
          {
            "description": "Anchor audit log entry ID",
            "name": "log_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Max entries in thread (default 40, max 200)",
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auditlog.ConversationResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/admin/api/v1/audit/log": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get paginated audit log entries",
        "parameters": [
          {
            "description": "Number of days (default 30)",
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Start date (YYYY-MM-DD)",
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "End date (YYYY-MM-DD)",
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by requested model selector",
            "name": "requested_model",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by provider name or provider type",
            "name": "provider",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by HTTP method",
            "name": "method",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by request path",
            "name": "path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by tracked user path subtree",
            "name": "user_path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by error type",
            "name": "error_type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by status code",
            "name": "status_code",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Filter by stream mode (true/false)",
            "name": "stream",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "Search across request_id/requested_model/provider/method/path/error_type/error_message",
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Page size (default 25, max 100)",
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Offset for pagination",
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/admin.auditLogListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/admin/api/v1/dashboard/config": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get dashboard runtime configuration",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/admin.DashboardConfigResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/admin/api/v1/models": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List all registered models with provider info and access state",
        "parameters": [
          {
            "description": "Filter by model category",
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/admin.modelInventoryResponse"
                  },
                  "maxItems": 10000,
                  "description": "Bounded by maxItems=10000."
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/admin/api/v1/models/categories": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List model categories with counts",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/providers.CategoryCount"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://gomodel.example.com",
      "description": "GoModel HTTPS deployment"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local GoModel"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "admin.DashboardConfigResponse": {
        "type": "object",
        "properties": {
          "BUDGETS_ENABLED": {
            "type": "string"
          },
          "CACHE_ENABLED": {
            "type": "string"
          },
          "FEATURE_FALLBACK_MODE": {
            "type": "string"
          },
          "GUARDRAILS_ENABLED": {
            "type": "string"
          },
          "LOGGING_ENABLED": {
            "type": "string"
          },
          "REDIS_URL": {
            "type": "string"
          },
          "SEMANTIC_CACHE_ENABLED": {
            "type": "string"
          },
          "USAGE_ENABLED": {
            "type": "string"
          },
          "USAGE_PRICING_RECALCULATION_ENABLED": {
            "type": "string"
          }
        }
      },
      "admin.auditLogEntryResponse": {
        "type": "object",
        "properties": {
          "alias_used": {
            "type": "boolean"
          },
          "auth_key_id": {
            "type": "string"
          },
          "auth_method": {
            "type": "string"
          },
          "cache_type": {
            "type": "string"
          },
          "client_ip": {
            "type": "string"
          },
          "data": {
            "description": "Data contains flexible request/response information as JSON",
            "allOf": [
              {
                "$ref": "#/components/schemas/auditlog.LogData"
              }
            ]
          },
          "duration_ns": {
            "description": "DurationNs is the request duration in nanoseconds",
            "type": "integer"
          },
          "error_type": {
            "type": "string"
          },
          "id": {
            "description": "ID is a unique identifier for this log entry (UUID)",
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "provider": {
            "description": "canonical provider type used for routing and filters",
            "type": "string"
          },
          "provider_name": {
            "type": "string"
          },
          "request_id": {
            "description": "Extracted fields for efficient filtering (indexed in relational DBs)",
            "type": "string"
          },
          "requested_model": {
            "description": "Core fields (indexed for queries)",
            "type": "string"
          },
          "resolved_model": {
            "type": "string"
          },
          "status_code": {
            "type": "integer"
          },
          "stream": {
            "type": "boolean"
          },
          "timestamp": {
            "description": "Timestamp is when the request started",
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/usage.RequestUsageSummary"
          },
          "user_path": {
            "type": "string"
          },
          "workflow_version_id": {
            "type": "string"
          }
        }
      },
      "admin.auditLogListResponse": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/admin.auditLogEntryResponse"
            }
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
...
      "admin.modelAccessResponse": {
        "type": "object",
        "properties": {
          "default_enabled": {
            "type": "boolean"
          },
          "effective_enabled": {
            "type": "boolean"
          },
          "override": {
            "$ref": "#/components/schemas/modeloverrides.Override"
          },
          "selector": {
            "type": "string"
          },
          "user_paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "admin.modelInventoryResponse": {
        "type": "object",
        "properties": {
          "access": {
            "$ref": "#/components/schemas/admin.modelAccessResponse"
          },
          "model": {
            "$ref": "#/components/schemas/core.Model"
          },
          "provider_name": {
            "type": "string"
          },
          "provider_type": {
            "type": "string"
          },
          "selector": {
            "type": "string"
          }
        }
      },
...
      "auditlog.ConversationResult": {
        "type": "object",
        "properties": {
          "anchor_id": {
            "type": "string"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/auditlog.LogEntry"
            }
          }
        }
      },
...
      "auditlog.LogData": {
        "type": "object",
        "properties": {
          "api_key_hash": {
            "type": "string"
          },
          "error_code": {
            "type": "string"
          },
          "error_message": {
            "description": "Error details (message can be long, so kept in JSON)",
            "type": "string"
          },
          "failover": {
            "description": "Failover captures runtime redirect details when translated execution\nmoved from the primary selector to a configured failover target.",
            "allOf": [
              {
                "$ref": "#/components/schemas/auditlog.FailoverSnapshot"
              }
            ]
          },
          "max_tokens": {
            "type": "integer"
          },
          "request_body": {
            "description": "Optional bodies (when LOGGING_LOG_BODIES=true)\nStored as interface{} so MongoDB serializes as native BSON documents (queryable/readable)\ninstead of BSON Binary (base64 in Compass)"
          },
          "request_body_too_big_to_handle": {
            "description": "Body capture status flags (set when body exceeds 1MB limit)",
            "type": "boolean"
          },
          "request_headers": {
            "description": "Optional headers (when LOGGING_LOG_HEADERS=true)\nSensitive headers are auto-redacted",
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "response_body": {},
          "response_body_too_big_to_handle": {
            "type": "boolean"
          },
          "response_headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "temperature": {
            "description": "Request parameters",
            "type": "number"
          },
          "user_agent": {
            "description": "Identity",
            "type": "string"
          },
          "workflow_features": {
            "description": "WorkflowFeatures captures the request-time effective workflow features\nafter runtime caps were applied. This keeps audit views historically accurate\neven if the active process config changes later.",
            "allOf": [
              {
                "$ref": "#/components/schemas/auditlog.WorkflowFeaturesSnapshot"
              }
            ]
          }
        }
      },
      "auditlog.LogEntry": {
        "type": "object",
        "properties": {
          "alias_used": {
            "type": "boolean"
          },
          "auth_key_id": {
            "type": "string"
          },
          "auth_method": {
            "type": "string"
          },
          "cache_type": {
            "type": "string"
          },
          "client_ip": {
            "type": "string"
          },
          "data": {
            "description": "Data contains flexible request/response information as JSON",
            "allOf": [
              {
                "$ref": "#/components/schemas/auditlog.LogData"
              }
            ]
          },
          "duration_ns": {
            "description": "DurationNs is the request duration in nanoseconds",
            "type": "integer"
          },
          "error_type": {
            "type": "string"
          },
          "id": {
            "description": "ID is a unique identifier for this log entry (UUID)",
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "provider": {
            "description": "canonical provider type used for routing and filters",
            "type": "string"
          },
          "provider_name": {
            "type": "string"
          },
          "request_id": {
            "description": "Extracted fields for efficient filtering (indexed in relational DBs)",
            "type": "string"
          },
          "requested_model": {
            "description": "Core fields (indexed for queries)",
            "type": "string"
          },
          "resolved_model": {
            "type": "string"
          },
          "status_code": {
            "type": "integer"
          },
          "stream": {
            "type": "boolean"
          },
          "timestamp": {
            "description": "Timestamp is when the request started",
            "type": "string"
          },
          "user_path": {
            "type": "string"
          },
          "workflow_version_id": {
            "type": "string"
          }
        }
      },
...
      "core.ErrorType": {
        "type": "string",
        "enum": [
          "provider_error",
          "rate_limit_error",
          "invalid_request_error",
          "authentication_error",
          "not_found_error"
        ],
        "x-enum-varnames": [
          "ErrorTypeProvider",
          "ErrorTypeRateLimit",
          "ErrorTypeInvalidRequest",
          "ErrorTypeAuthentication",
          "ErrorTypeNotFound"
        ]
      },
...
      "core.GatewayError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string"
          },
          "param": {
            "type": "string",
            "nullable": true
          },
          "provider": {
            "type": "string"
          },
          "status_code": {
            "type": "integer"
          },
          "type": {
            "$ref": "#/components/schemas/core.ErrorType"
          }
        }
      },
...
      "core.Model": {
        "type": "object",
        "properties": {
          "created": {
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "metadata": {
            "description": "Metadata holds optional enrichment data (display name, pricing, capabilities, etc.).\nMay be nil if the model was not found in the external registry.",
            "allOf": [
              {
                "$ref": "#/components/schemas/core.ModelMetadata"
              }
            ]
          },
          "object": {
            "type": "string"
          },
          "owned_by": {
            "type": "string"
          }
        }
      },
      "core.ModelCategory": {
        "type": "string",
        "enum": [
          "all",
          "text_generation",
          "embedding",
          "image",
          "audio",
          "video",
          "utility"
        ],
        "x-enum-varnames": [
          "CategoryAll",
          "CategoryTextGeneration",
          "CategoryEmbedding",
          "CategoryImage",
          "CategoryAudio",
          "CategoryVideo",
          "CategoryUtility"
        ]
      },
...
      "modeloverrides.Override": {
        "type": "object",
        "properties": {
          "created_at": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "provider_name": {
            "type": "string"
          },
          "selector": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "user_paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "providers.CategoryCount": {
        "type": "object",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/core.ModelCategory"
          },
          "count": {
            "type": "integer"
          },
          "display_name": {
            "type": "string"
          }
        }
      },
...
      "usage.RequestUsageSummary": {
        "type": "object",
        "properties": {
          "cache_write_input_tokens": {
            "type": "integer"
          },
          "cached_input_ratio": {
            "type": "number"
          },
          "cached_input_tokens": {
            "type": "integer"
          },
          "entries": {
            "type": "integer"
          },
          "estimated_cached_characters": {
            "type": "integer"
          },
          "input_tokens": {
            "type": "integer"
          },
          "output_tokens": {
            "type": "integer"
          },
          "total_tokens": {
            "type": "integer"
          },
          "uncached_input_tokens": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}